integrated sqldelight and koin

+++ common/src/commonMain/kotlin/dev/baseio/slackclone/domain/usecases/channels/UseCaseFetchChannelCount.kt
@@ -4,8 +4,7 @@ import dev.baseio.slackclone.domain.repository.ChannelsRepository
 import dev.baseio.slackclone.domain.usecases.BaseUseCase
 
 class UseCaseFetchChannelCount(private val channelsRepository: ChannelsRepository) : BaseUseCase<Int,Unit> {
-
   override suspend fun perform(): Int {
-    return channelsRepository.channelCount()
+    return channelsRepository.channelCount().toInt()
   }
 }
